requirements.txt
flask_cors
flask
psycopg2-binary
flask_sqlalchemy
pyinstaller